move community checklist to a modal
