+++ css/posts.less
@@ -1114,6 +1114,9 @@ ul.posts {
   padding: 25px 25px 20px;
   background-color: white;
   margin-bottom: 9px;
+  a {
+    color: @damian-green;
+  }
 }
 
 .post-list-footer {
